fix conflicts

+++ components/raftstore/src/store/peer_storage.rs
@@ -110,8 +110,8 @@ pub fn last_index(state: &RaftLocalState) -> u64 {
 struct EntryCache {
     cache: VecDeque<Entry>,
     trace: VecDeque<CachedEntries>,
-    hit: Cell<i64>,
-    miss: Cell<i64>,
+    hit: Cell<u64>,
+    miss: Cell<u64>,
     #[cfg(test)]
     size_change_cb: Option<Box<dyn Fn(i64) + Send + 'static>>,
 }
@@ -1226,7 +1226,7 @@ where
             let drain_to = if half { cache_len / 2 } else { cache_len - 1 };
             let idx = cache.cache[drain_to].index;
             let mem_size_change = cache.compact_to(idx + 1);
-            RAFT_ENTRIES_EVICT_BYTES.inc_by(mem_size_change as i64);
+            RAFT_ENTRIES_EVICT_BYTES.inc_by(mem_size_change);
         }
     }
 
@@ -1245,8 +1245,8 @@ where
         }
         if let Some(stats) = self.engines.raft.flush_stats() {
             RAFT_ENTRIES_CACHES_GAUGE.set(stats.cache_size as i64);
-            RAFT_ENTRY_FETCHES.hit.inc_by(stats.hit as i64);
-            RAFT_ENTRY_FETCHES.miss.inc_by(stats.miss as i64);
+            RAFT_ENTRY_FETCHES.hit.inc_by(stats.hit as u64);
+            RAFT_ENTRY_FETCHES.miss.inc_by(stats.miss as u64);
         }
     }
